19: 修订分布式训练命名
