+++ easytext/metrics/acc_metric.py
@@ -13,6 +13,7 @@ Date:    2020/05/30 07:36:00
 from typing import Dict
 
 import torch
+from torch import distributed as TorchDist
 
 from easytext.metrics import Metric
 
@@ -77,10 +78,10 @@ class AccMetric(Metric):
     def _distributed_data(self):
         distributed_tensor = self._data.to_tensor()
 
-        if torch.distributed.get_backend() == "nccl":
+        if TorchDist.get_backend() == "nccl":
             distributed_tensor.to(torch.distributed.get_rank())
 
-        torch.distributed.all_reduce(tensor=distributed_tensor)
+        TorchDist.all_reduce(tensor=distributed_tensor)
 
         self._data.update_from_tensor(distributed_tensor)
