+++ easytext/trainer/trainer.py
@@ -14,12 +14,15 @@ Date:    2020/05/16 00:34:00
 import os
 import torch
 import logging
-from typing import List, Union
 from tqdm import tqdm
 import shutil
 
+from typing import Optional
+
 from torch.utils.data import DataLoader
 from torch.optim.lr_scheduler import ReduceLROnPlateau
+from torch.nn.parallel import DistributedDataParallel
+from torch import distributed as TorchDist
 
 from easytext.model import Model
 from easytext.loss import Loss
@@ -29,13 +32,15 @@ from easytext.data.model_collate import ModelInputs
 from easytext.metrics import ModelMetricAdapter
 from easytext.utils.json_util import json2str
 from easytext.utils.nn import cuda_util
+from easytext.utils.distributed_util import DistributedFuncWrapper
 from easytext.trainer.metric_tracker import MetricTracker
 from easytext.trainer.grad_rescaled import GradRescaled
 from easytext.trainer import Record
 from easytext.trainer.trainer_callback import TrainerCallback
+from easytext.distributed import Distributed
 
 
-class Trainer(TrainerCallback):
+class Trainer(TrainerCallback, Distributed):
     """
     训练器
     """
@@ -49,12 +54,14 @@ class Trainer(TrainerCallback):
                  loss: Loss,
                  metrics: ModelMetricAdapter,
                  optimizer_factory: OptimizerFactory,
+                 device: torch.device,
+                 is_distributed: bool = False,
                  lr_scheduler_factory: LRSchedulerFactory = None,
                  grad_scaled: GradRescaled = None,
                  patient: int = None,
                  num_check_point_keep: int = None,
-                 devices: Union[str, List[str]] = None,
-                 trainer_callback: TrainerCallback = None):
+                 trainer_callback: TrainerCallback = None,
+                 ):
         """
         训练器初始化
         :param num_epoch: 训练的 epoch 数量
@@ -64,41 +71,19 @@ class Trainer(TrainerCallback):
         :param optimizer_factory: 模型的优化器的创建工厂。为什么不直接使用优化器？是因为, 优化器的创建依赖于 model, 所以
         这里传递的参数 optimizer factory, 避免使用者在 trainer 外面生成 optimizer, 导致在 trainer 内 optimizer 依赖于
         model 的参数产生问题。典型问题是: 设置 cuda.
+        :param device: 训练时所依赖的 device
+        :param is_distributed: 当前 trainer 是否是在多 GPU 环境下使用
         :param serialize_dir: 训练存储的文件路径
         :param patient: early stopping 的 patient. 如果是 `None`, 将不会进行 early stopping;
         否则, 当前训练的指标超出了 patient 个 epoch 将会 early stopping.
         :param num_check_point_keep: checkpoint 保留的数量。如果是 `None` 则全部保留;
         否则，保留 num_check_point_keep 个checkpoint.
-        :param devices: device 字符串, "cuda:0" 或者 "cpu"; 列表类型，多个gpu,  例如 ["cuda:0", "cuda:1"]; 如果是 cpu, 只能是 ["cpu"].
         """
-        if devices is None or len(devices) == 0:
-            self._devices = [torch.device("cpu")]
-        else:
 
-            if isinstance(devices, str):
-                self._devices = [torch.device(devices)]
-            elif isinstance(devices, list):
-                self._devices = [torch.device(device) for device in devices]
-            else:
-                raise RuntimeError(f"devices type: {type(devices)} 不是 str 或者 list!")
-
-            if len(self._devices) != 1:
-                raise RuntimeError(f"目前仅仅支持单 GPU 或 CPU 训练, 设置的 cuda devices 是 {devices}")
-
-        if self._devices[0].type == "cuda":
-            self._model = model.cuda(self._devices[0])
-        else:
-            self._model = model.cpu()
-
+        self._device = device
         self._loss = loss
         self._metrics = metrics
-        self._optimizer = optimizer_factory.create(model=self._model)
-
-        if lr_scheduler_factory is not None:
-            self._lr_scheduler = lr_scheduler_factory.create(optimizer=self.optimizer,
-                                                             model=self.model)
-        else:
-            self._lr_scheduler = None
+        self._optimizer_factory = optimizer_factory
 
         self._grad_scaled = grad_scaled
 
@@ -106,8 +91,41 @@ class Trainer(TrainerCallback):
         self._metric_tracker = MetricTracker(patient=patient)
         self._num_check_point_keep = num_check_point_keep
         self._num_epoch = num_epoch
-        self._current_epoch: int = None
+        self._current_epoch: Optional[int] = None
         self._trainer_callback = trainer_callback
+
+        self._is_distributed = is_distributed
+
+        if self._is_distributed:
+            self._distributed_func_wrapper = DistributedFuncWrapper(dst_rank=0)
+        else:
+            self._distributed_func_wrapper = DistributedFuncWrapper(dst_rank=None)
+
+        self._model = model
+        if self._is_distributed:
+
+            assert self._device.type != "cpu", f"多 GPU 训练, device 不能是 cpu"
+
+            torch.cuda.set_device(self._device)
+            self._model.cuda(self._device)
+            self._optimizer = optimizer_factory.create(self._model)
+
+            self._model = DistributedDataParallel(module=self._model,
+                                                  device_ids=[self._device],
+                                                  output_device=self._device)
+        else:
+            self._model = self._model.to(self._device)
+            self._optimizer = self._optimizer_factory.create(self._model)
+
+        if lr_scheduler_factory is not None:
+            self._lr_scheduler = lr_scheduler_factory.create(optimizer=self.optimizer,
+                                                             model=self.model)
+        else:
+            self._lr_scheduler = None
+
+    @property
+    def is_distributed(self) -> bool:
+        return self._is_distributed
 
     @property
     def model(self):
@@ -250,13 +268,15 @@ class Trainer(TrainerCallback):
 
         if last_epoch is not None:
             self._current_epoch = last_epoch
-            logging.info(f"Load checkpoint, 当前 epoch: {last_epoch}")
+
+            self._distributed_func_wrapper(logging.info, f"Load checkpoint, 当前 epoch: {last_epoch}")
+
             saved_dir = os.path.join(serialize_dir, f"checkpoint_epoch_{last_epoch}")
 
             model_file_path = os.path.join(saved_dir, "model.pt")
             self._model.load_state_dict(torch.load(model_file_path))
 
-            logging.info(f"last epoch{last_epoch}, loaded: {self._model.state_dict()}")
+            self._distributed_func_wrapper(logging.info, f"last epoch{last_epoch}, loaded: {self._model.state_dict()}")
 
             optimizer_file_path = os.path.join(saved_dir, "optimizer.pt")
             self._optimizer.load_state_dict(torch.load(optimizer_file_path))
@@ -300,9 +320,9 @@ class Trainer(TrainerCallback):
                                                    model_inputs.labels
 
                 # 设置到 cuda 训练
-                if self._devices[0].type == "cuda":  # 仅仅处理 GPU, 默认使用 CPU
-                    batch_inputs = cuda_util.cuda(batch_inputs, cuda_device=self._devices[0])
-                    labels = cuda_util.cuda(labels, cuda_device=self._devices[0])
+                if self._device.type == "cuda":  # 仅仅处理 GPU, 默认使用 CPU
+                    batch_inputs = cuda_util.cuda(batch_inputs, cuda_device=self._device)
+                    labels = cuda_util.cuda(labels, cuda_device=self._device)
 
                 outputs = self._model(**batch_inputs)
                 batch_loss: torch.Tensor = self._loss(outputs, labels)
@@ -321,12 +341,14 @@ class Trainer(TrainerCallback):
                 total_num += batch_size
 
                 batch_metrics, target_metric = self._metrics(model_outputs=outputs, golden_labels=labels)
-                logging.info(f"Epoch: {self._current_epoch}, batch loss: {batch_loss},"
-                             f"batch metrics: {json2str(batch_metrics)}, "
-                             f"target metric: {json2str(target_metric)}")
+                self._distributed_func_wrapper(
+                    logging.info,
+                    f"Epoch: {self._current_epoch}, batch loss: {batch_loss},"
+                    f"batch metrics: {json2str(batch_metrics)}, "
+                    f"target metric: {json2str(target_metric)}")
 
         # total_loss = total_loss / total_num 这是合理的 loss, 因为所有的 total_num 是一样的所以，没有必要再除以一次了
-        return total_loss
+        return total_loss, total_num
 
     def recovery_train(self,
                        train_data_loader: DataLoader,
@@ -367,13 +389,26 @@ class Trainer(TrainerCallback):
                 break
         return is_empty
 
-    def train(self, train_data_loader: DataLoader,
-              validation_data_loader: DataLoader) -> None:
+
+
+    def train(self,
+               train_data_loader: DataLoader,
+               validation_data_loader: DataLoader) -> None:
+        """
+        模型训练
+        :param train_data_loader: 训练集 data loader
+        :param validation_data_loader: 验证集 data loader
+        :return:
+        """
+
         if not self._is_serialize_empty():
             raise RuntimeError(f"新训练，请清空保存文件件: {self._serialize_dir}")
 
-        return self._train(train_data_loader=train_data_loader,
-                           validation_data_loader=validation_data_loader)
+        if self._is_distributed:
+            TorchDist.barrier()
+
+        self._train(train_data_loader=train_data_loader,
+                    validation_data_loader=validation_data_loader)
 
     def _train(self,
                train_data_loader: DataLoader,
@@ -397,12 +432,28 @@ class Trainer(TrainerCallback):
 
             self._current_epoch = epoch
 
-            logging.info(f"Start train epoch: {self._current_epoch}")
+            self._distributed_func_wrapper(logging.info, f"Start train epoch: {self._current_epoch}")
+
             self.on_train_epoch_start(trainer=self, record=record)
 
-            train_loss = self._train_or_evaluate(phrase=Trainer._TRAIN,
+            train_loss, total_num = self._train_or_evaluate(phrase=Trainer._TRAIN,
                                                  data_loader=train_data_loader)
-            record.epoch_train_loss = train_loss
+
+            if self._is_distributed:
+                # 对于分布式来说需要得到分布式的 loss
+                dist_loss_tensor = torch.tensor([train_loss, total_num], dtype=torch.float)
+
+                if TorchDist.get_backend() == "nccl":
+                    dist_loss_tensor.to(TorchDist.get_rank())
+
+                TorchDist.all_reduce(dist_loss_tensor)
+
+                record.epoch_train_loss = dist_loss_tensor[0].item()
+                record.epoch_train_num = int(dist_loss_tensor[1].item())
+            else:
+
+                record.epoch_train_loss = train_loss
+                record.epoch_train_num = total_num
 
             # 输出metrics
             train_metric_dict, train_target_metric = self._metrics.metric
@@ -410,16 +461,33 @@ class Trainer(TrainerCallback):
             record.train_metric = train_metric_dict
             record.train_target_metric = train_target_metric
 
-            logging.info(
-                f"Train epoch: {epoch}, loss: {train_loss}, target metric: {train_target_metric.name}:{train_target_metric.value} "
-                f"metrics: {json2str(train_metric_dict)}")
+            self._distributed_func_wrapper(logging.info,
+                                           f"Train epoch: {epoch}, "
+                                           f"loss: {train_loss}, "
+                                           f"target metric: {train_target_metric.name}:{train_target_metric.value},"
+                                           f"metrics: {json2str(train_metric_dict)}")
 
             self.on_train_epoch_stop(trainer=self, record=record)
 
             self.on_evaluate_validation_epoch_start(trainer=self, record=record)
 
-            validation_loss = self.evaluate(validation_data_loader=validation_data_loader)
-            record.epoch_validation_loss = validation_loss
+            validation_loss, validation_num = self.evaluate(validation_data_loader=validation_data_loader)
+
+            if self._is_distributed:
+                # 对于分布式来说需要得到分布式的 loss
+                dist_loss_tensor = torch.tensor([validation_loss, validation_num], dtype=torch.float)
+
+                if TorchDist.get_backend() == "nccl":
+                    dist_loss_tensor.to(TorchDist.get_rank())
+
+                TorchDist.all_reduce(dist_loss_tensor)
+
+                record.epoch_validation_loss = dist_loss_tensor[0].item()
+                record.epoch_validation_num = int(dist_loss_tensor[1].item())
+            else:
+
+                record.epoch_validation_loss = validation_loss
+                record.epoch_validation_num = validation_num
 
             validation_metric_dict, validation_target_metric = self._metrics.metric
             record.validation_metric = validation_metric_dict
@@ -430,7 +498,9 @@ class Trainer(TrainerCallback):
                                             train_model_target_metric=train_target_metric,
                                             validation_metric=validation_metric_dict,
                                             validation_model_target_metric=validation_target_metric)
-            logging.info(
+
+            self._distributed_func_wrapper(
+                logging.info,
                 f"Evaluate Valid epoch: {epoch}, loss: {validation_loss}, "
                 f"target metric: {validation_target_metric.name}:{validation_target_metric.value} "
                 f"metrics: {json2str(validation_metric_dict)}")
@@ -448,40 +518,44 @@ class Trainer(TrainerCallback):
                 else:
                     self._lr_scheduler.step(epoch=epoch)
 
-            self.save_checkpoint(epoch=epoch)
+            self._distributed_func_wrapper(self.save_checkpoint, epoch=epoch)
 
             if self._metric_tracker.early_stopping(epoch):
-                logging.info(f"Epoch: {epoch}, early stopping!")
+                self._distributed_func_wrapper(
+                    logging.info, f"Epoch: {epoch}, early stopping!")
                 break
 
         self.on_training_complete(trainer=self, record=record)
 
     def on_train_epoch_start(self, trainer: "Trainer", record: Record) -> None:
-        logging.info(f"on_train_epoch_start: {record.epoch}")
+
+        self._distributed_func_wrapper(logging.info, f"on_train_epoch_start: {record.epoch}")
+
         if self._trainer_callback is not None:
             self._trainer_callback.on_train_epoch_start(trainer=trainer,
                                                         record=record)
 
     def on_train_epoch_stop(self, trainer: "Trainer", record: Record) -> None:
-        logging.info(f"on_train_epoch_stop: {record.epoch}")
+        self._distributed_func_wrapper(logging.info, f"on_train_epoch_stop: {record.epoch}")
+
         if self._trainer_callback is not None:
             self._trainer_callback.on_train_epoch_stop(trainer=trainer,
                                                        record=record)
 
     def on_evaluate_validation_epoch_start(self, trainer: "Trainer", record: Record) -> None:
-        logging.info(f"on_evaluate_epoch_start: {record.epoch}")
+        self._distributed_func_wrapper(logging.info, f"on_evaluate_epoch_start: {record.epoch}")
         if self._trainer_callback is not None:
             self._trainer_callback.on_evaluate_validation_epoch_start(trainer=trainer,
                                                                       record=record)
 
     def on_evaluate_validation_epoch_stop(self, trainer: "Trainer", record: Record) -> None:
-        logging.info(f"on_evaluate_epoch_stop: {record.epoch}")
+        self._distributed_func_wrapper(logging.info, f"on_evaluate_epoch_stop: {record.epoch}")
         if self._trainer_callback is not None:
             self._trainer_callback.on_evaluate_validation_epoch_stop(trainer=trainer,
                                                                      record=record)
 
     def on_training_complete(self, trainer: "Trainer", record: Record) -> None:
-        logging.info(f"on_training_complete: {record.epoch}")
+        self._distributed_func_wrapper(logging.info, f"on_training_complete: {record.epoch}")
         if self._trainer_callback is not None:
             self._trainer_callback.on_training_complete(trainer=trainer,
                                                         record=record)
